Corrected LocalDate Format and add Swipe Helper class
