+++ app/build.gradle
@@ -96,5 +96,5 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
     // Extended Floating Action Button
-    implementation "com.google.android.material:material:1.4.0-alpha01"
+    implementation "com.google.android.material:material:1.4.0-alpha02"
 }
